Feat. Finding user name with our token
We know that out token has a payload, that payload contains an userId, with that we created a new function in our web socket service that finds the user, otherwise we sent an Error and disconnected him from the service.
Now we are able to capture the user full name and send it to the client.
It's working.